More refactorings!
